requirements.txt
albumentations==0.4.6
efficientnet-pytorch==0.7.0
catalyst==20.8.2
graphql-core==2.0
joblib==0.16.0
kaggle==1.5.6
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
pillow==7.2.0
seaborn==0.10.1
scikit-learn==0.23.2
torch==1.6.0
torchvision==0.7.0
wandb==0.9.6